refactor: Rename core -> lib refactor: Rename domain -> core refactor: Remove lib/internal, core/internal
