+++ src/components/BottomNav/BottomNav.ts
@@ -1,4 +1,4 @@
-import { Component, Watch, Vue } from 'vue-property-decorator';
+import { Component, Vue } from 'vue-property-decorator';
 import {
   domainStreamEditorModule,
   uiBottomNavModule,
