+++ src/core/ColorDefinition.ts
