+++ src/core/StreamEvent.ts
