--- src/core/internal.ts
@@ -1,1 +1,0 @@
-export * from './InstanceMap';
