--- src/domain/internal.ts
@@ -1,2 +1,0 @@
-export * from './StreamEvent';
-export * from './StreamDataset';
