+++ src/lib/InstanceMap.ts
